2 лаба понизил версию платформы и студии

+++ lab2/1.vector/1.vector/1.vector.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-	vector<float> inputArray;
+	vector<double> inputArray;
 	try
 	{
 		FillArray(cin, inputArray);
