+++ lab2/1.vector/1.vector/arrayProcessor.h
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <vector>
 
-void FillArray(std::istream& input, std::vector<float>& array);
-void MultiplyOnMinimal(std::vector<float>& array);
-void DisplayArray(std::ostream& output, std::vector<float>& array);
-void SortArray(std::vector<float>& array);
+void FillArray(std::istream& input, std::vector<double>& array);
+void MultiplyOnMinimal(std::vector<double>& array);
+void DisplayArray(std::ostream& output, std::vector<double>& array);
+void SortArray(std::vector<double>& array);
